requirements.txt
numpy==1.25.0
scikit-learn==1.2.1
opencv_python==4.7.0.72
opencv-python-headless==4.8.0.74
Pillow==8.3.1
streamlit==1.24.0
tensorflow==2.13.0rc1
